helper component added to hold form validations and the CSS
