+++ src/app/components/login/login.component.html
@@ -9,37 +9,32 @@
                     <div class="panel-body p-3">
                         <form [formGroup]="loginForm">
                             <div class="form-group pt-3">
-                                <div class="input-field">
+                                <div class="input-field" [class.error]="loginForm.controls['email'].dirty && loginForm.controls['email'].hasError('required')">
                                     <label for="exampleInputEmail1">Email address</label>
                                     <input formControlName="email" type="email" class="form-control" id="exampleInputEmail1"
                                         aria-describedby="emailHelp" placeholder="Enter email">
                                 </div>
+                                <small *ngIf="loginForm.controls['email'].dirty && loginForm.controls['email'].hasError('required')" class="text-danger">* Email is required</small>
                             </div>
-                            <small *ngIf="loginForm.controls['email'].dirty && loginForm.hasError('required', 'email')" class="text-danger">*email required</small>
-                                    <div class="form-group pt-3">
-                                        <div class="input-field">
-                                            <label for="exampleInputPassword1">Password</label>
-                                            <input formControlName="password" type="password" class="form-control" id="exampleInputPassword1"
-                                                placeholder="Password">
-                                        </div>
-                                    </div>
-                                    <small *ngIf="loginForm.controls['password'].dirty && loginForm.hasError('required', 'password')" class="text-danger">*password required</small>
-                                     <div class="form-group pt-5">
-                                        <button (click)="onSubmit()" class="btn btn-success pt-6">Login</button>
-
-                                    </div>
-                                        <div class="form-inline fr mt-4">
-                                            <a id=" forgot" class="font-weight-lighter">Forgot your password?</a>
-                                        </div>
-                                        <div class="pt-4 text muted" >Don't have account? <a id="forget" class="font-weight-lighter" routerLink="/register"> Register</a></div>
-                                    
-                                
-                            
+                            <div class="form-group pt-3">
+                                <div class="input-field" [class.error]="loginForm.controls['password'].dirty && loginForm.controls['password'].hasError('required')">
+                                    <label for="exampleInputPassword1">Password</label>
+                                    <input formControlName="password" type="password" class="form-control" id="exampleInputPassword1"
+                                        placeholder="Password">
+                                </div>
+                                <small *ngIf="loginForm.controls['password'].dirty && loginForm.controls['password'].hasError('required')" class="text-danger">* Password is required</small>
+                            </div>
+                            <div class="form-group pt-5">
+                                <button (click)="onSubmit()" class="btn btn-success pt-6">Login</button>
+                            </div>
+                            <div class="form-inline fr mt-4">
+                                <a id="forgot" class="font-weight-lighter">Forgot your password?</a>
+                            </div>
+                            <div class="pt-4 text-muted">Don't have an account? <a id="forget" class="font-weight-lighter" routerLink="/register"> Register</a></div>
                         </form>
-                    
+                    </div>
                 </div>
             </div>
-
         </div>
     </div>
 </div>
